+++ react/src/components/Badge/index.js
@@ -1,0 +1,1 @@
+export { Badge, BadgePin, default } from './Badge';